# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Café Menu API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface AdminUser {
  email: string;
}

export interface LoginResult {
  success: boolean;
  user: AdminUser;
}

export interface Category {
  id: number;
  name: string;
  sortOrder: number;
}

export interface CategoryInput {
  name: string;
  sortOrder?: number;
}

export interface MenuItem {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  imageUrl?: string | null;
  categoryId: number;
  /** @nullable */
  categoryName?: string | null;
  available: boolean;
  sortOrder: number;
}

export interface MenuItemInput {
  name: string;
  description?: string;
  price: number;
  imageUrl?: string;
  categoryId: number;
  available?: boolean;
  sortOrder?: number;
}

export interface MenuItemUpdate {
  name?: string;
  description?: string;
  price?: number;
  imageUrl?: string;
  categoryId?: number;
  available?: boolean;
  sortOrder?: number;
}

export interface MenuSummary {
  totalItems: number;
  totalCategories: number;
  availableItems: number;
}

export type ListMenuItemsParams = {
  categoryId?: number;
  available?: boolean;
};