# lib/api-zod/src/generated/types/loginResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Café Menu API
 * OpenAPI spec version: 0.1.0
 */
import type { AdminUser } from "./adminUser";

export interface LoginResult {
  success: boolean;
  user: AdminUser;
}