# lib/api-zod/src/generated/types/menuItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Café Menu API
 * OpenAPI spec version: 0.1.0
 */

export interface MenuItem {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  imageUrl?: string | null;
  categoryId: number;
  /** @nullable */
  categoryName?: string | null;
  available: boolean;
  sortOrder: number;
}