# lib/api-zod/src/generated/types/menuItemUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Café Menu API
 * OpenAPI spec version: 0.1.0
 */

export interface MenuItemUpdate {
  name?: string;
  description?: string;
  price?: number;
  imageUrl?: string;
  categoryId?: number;
  available?: boolean;
  sortOrder?: number;
}