# lib/api-zod/src/generated/types/menuSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Café Menu API
 * OpenAPI spec version: 0.1.0
 */

export interface MenuSummary {
  totalItems: number;
  totalCategories: number;
  availableItems: number;
}